# spiffworkflow-backend/src/spiffworkflow_backend/api.yml
openapi: "3.0.2"
info:
  version: 1.0.0
  title: spiffworkflow-backend
  license:
    name: LGPL
servers:
  - url: http://localhost:5000/v1.0
# this is handled in flask now
security: []
paths:
...
  /login_return:
    parameters:
      - name: code
        in: query
        required: true
        schema:
          type: string
      - name: state
        in: query
        required: true
        schema:
          type: string
      - name: session_state
        in: query
        required: false
        schema:
          type: string
    get:
      operationId: spiffworkflow_backend.routes.authentication_controller.login_return
      tags:
        - Authentication
      responses:
        "200":
          description: Test Return Response
...
  /login_with_access_token:
    parameters:
      - name: access_token
        in: query
        required: true
        schema:
          type: string
      - name: authentication_identifier
        in: query
        required: true
        schema:
          type: string
    post:
      operationId: spiffworkflow_backend.routes.authentication_controller.login_with_access_token
      summary: Authenticate user for API access with an openid token already posessed.
      tags:
        - Authentication
      responses:
        "200":
          description: "Returns ok: true if user successfully logged in."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkTrue"
  /login_api:
    parameters:
      - name: authentication_identifier
        in: query
        required: true
        schema:
          type: string
    get:
      operationId: spiffworkflow_backend.routes.authentication_controller.login_api
      summary: Authenticate user for API access
      tags:
        - Authentication
      responses:
        "200":
          description: Redirects to authentication server
  /login_api_return:
    parameters:
      - name: code
        in: query
        required: true
        schema:
          type: string
      - name: state
        in: query
        required: true
        schema:
          type: string
      - name: session_state
        in: query
        required: false
        schema:
          type: string
    get:
      operationId: spiffworkflow_backend.routes.authentication_controller.login_api_return
      tags:
        - Authentication
      responses:
        "200":
          description: Test Return Response
...
components:
  securitySchemes:
    oAuth2AuthCode:
      type: oauth2
      description: authenticate with openid server
      flows:
        authorizationCode:
          authorizationUrl: /v1.0/login_api
          tokenUrl: /v1.0/login_api_return
          scopes:
            read_email: read email
      x-tokenInfoFunc: spiffworkflow_backend.routes.authentication_controller.get_scope
  schemas:
    OkTrue:
      properties:
        ok:
          type: boolean
          example: true
...